Move debug target path rendering to SelectionDecorations
